Move university factories
Factories should be located in the tests/ module of each app.
