+++ universities/tests/factories.py
@@ -4,7 +4,7 @@ import factory
 
 from django.core.files.uploadedfile import SimpleUploadedFile
 
-from .models import University
+from universities.models import University
 
 # base64 encoded png FUN logo
 FUN_LOGO = """iVBORw0KGgoAAAANSUhEUgAAAGQAAABDCAIAAADlKTeNAAAACXBIWXMAAAsTAAALEwEAmpwYAAAAIGNIUk0AAHolAACAgwAA+f8AAIDpAAB1MAAA6mAAADqYAAAXb5JfxUYAAA9USURBVHja3Jx5dFzVfcd/v3vfOrtGo9FolywZbLwLm2AWsyVAWQqBA2FpCQGThuJAU4pzXDiNIenC1kAhrA6FE7ZAEiiFAG6CMbbAGLBiW7ZlWbK1WuPRaPb1Lff2jzGOrM2WNZbGuef9NfPezH2f+3vf33Lvfcg5h4JviYwxGMvEMyZjAMA5giISj10utstT2Q2hANGkNWPrvkhnINXuT3T0J0MxPZbW4mkjo5kmBwBABFFAmyK6LKLXJS+oc8yrdZ47x0sJOa4dw8KxrM1tg837Ipt2D+71p3qCqVjKIAQlipJIFZFIIhUIIgAgAAfGuWawtMbSWSOjm7JMZ5ZaL1joven8mtoS218srI+3D6xrOdDUGuroT0SSukUSnKpgt4gCRYKYQ8NhlG4iACIiAudgMj4Q0/rD6RKn+PeX1N9z5ay/KFitvbF3vtj/h+bAzp54SjPcNrnYLkkCAQTOj6VTBJFSHIxpu3qjN5xT/eKdSwjBEx7Wh83+1zf2bGgJDsSyHofsdcoSJYwDy0dPKEEA2LQ7dM2ZFa/c/Y0TGNZvP+t58Y9dn+0OEcSqYotdFUzOGctzBxCBEmxqDa6+bvZ918458WC9/XnfMx/s+2z3gF2VaoqtgoCmyY7fHxNEzWDbu6Nr7z/7rNmeEwbWxl0Dj77T9sc/DThVqdZrIQQMcyqGR6Skw5+o99k++tmyfInXcYQVjGfvf23H6xt6CSEzfTZKpwjTEF74xZ7Q07c33nR+bUHDevPT3gde39kZSM6ucFgVQTPY1DtcitgXSlcUKx/97ByB5iFezX/Im8zqf/fUVzc/9kVGY4sb3JJIpoUUAJicl7vV7V3x97f050cK89u/rfvC5927/tX13QvqXOVuRdPZ9Ma8BEERyG8+7Ss4WK+s7/zmTz7ZH8ouaXATmGqFGt24GJQXK5t2D/YEUwUE66G3di1/ckup03JSuS1rsAJJODlwqyIGItr/bT1QKLDue3X7fS/vnFvt9NilrM6gkBrn3CLRpp3BgoB136stD/5296kz3KpEdbOwSOWMq8gm7eyNxdPaNMP6z//Z/R9vti6pdwsCmKwQ64icg8Mq9gTSzXuj0wnrN029q37Vsqi+SKCEMSjYRhEyutnam5g2WG37Y3c892VDmd0qCYVpU0Njb1GgWzvD0wbrrue3AVCvQy5AnRrZVIl2D6SmB9YLf+hc3zIwq8I+XdH5RDXepgjBmBaMZ6caVipr/OK99uoSlZ0AE0MHNV6VaDCW7fQnpxrW6xt69vgTpS6FnSi0ACjFRMYMxCZlWccyFfbul36XRRwJChEB+PgjfPDMg0/H2IKMMDSpzE3qHDHNxLHPIQCmySJJbUphte2PtXRGfIebFSFomnx/KGVyGLUWYhhMkYUyl8o4o4gHollK0OOQRuaPlGBaM8NxvcQpCwIyxgWKobhuMO51SmPlmwLFeNqIp41Spzw6VgREHIhNLawNO4LBuOZzqeaQHpkmT2vGmbPdVkXUdAaHFyY54zZV8EfS7fuTTqscT+t1XjVrsGjSsMh0mIUaJiME51TbekNpQgSCmMgYXqeIiH2hbKlL1ke4FETI6kyViM9l8Ud0q0z5KFaLBCE0xbA2tYVlkfLDyiDYGUxetKhkzYrTxr/23HvXtfXHK4vUN358+kBMu/SBTwmiJJKhFZV2f/JHV8y856qT7/7l1jeb+iqL1UhSX7Oisb7MfvaqdZ2BZJ3XOswFcw5dA6nHls+/6vSKax/5fNu+aHmROqw2hABAMJrSpw6WbrA9/XGnRRg2cJrOqkusABBJanv2J1SJjnDe0D2QyugcOXocsl2VREGwKYJmmENhMYCMyep8NgSs8lgyOssazGUVKzwWhyo+/YPGy37aFEnqdvWwMJhzQIDqEoso0IpidVNraIzaFqY1c+pg7eiO7R/MuKzSsGI0QchoDAA+3HLg+oc3zaywj9T+jG76nHKRVcyaDAASGZ0xlpuQH6ZZqYwBAGnNpAQQkHGeSOteh3LGLM8jt8y/45nmxhnu4R4AIZUxASCrMzrG9AQB0CdXYpsYrK5AIprUSxxjrl2RROKySU6rONIPum2SQNFIGccw00IIYYwTgrdcULelPfLSR12LG1yaMbE7R4RJxjoTi7N6QxnOAce+XYIoCUSkww9JIIjHPudsU4RERnvorVYAePy2hYtnulr74rJAJxjHwyTDQjJBy0oKNLdaY9wxPPzIS7Op0kO/a3vq/XZK8IUfLqaI/eG0QHFCtCaId3KwgnGNjjthmdHNQCwzEM8eOvzRjGYwOrlpTs45Qaz3We98tnlHd7Su1PbcisaewVRGY0f/w4yDOLkJsYldnNZGkeQhZg5OizC3yl5fZv36sMyvdSJiXvLtEocii8KtT3yZ1cxLTi1fedXJ27uiBMlRE2c2eVKL9yZ2sWFwHHckL24su7ixbNiHr6zv+oc1fzq53DFJWJph1pfZdvclvv/0lpfuWvIv3zmleW9kw47g3BrnEdWIA3AOLps4hYk08nGVChIZY++BZE8wlTt6g6nugXTXQFKVaV6US9PZvFrXGxt7n3i3HQBe+OHiUpfUHUgJwhFvhJsMPE556ixLoGT8Ifz9V/3f/fnm+jL70ITZ5LymxMYYRwD42pkyBmyEY82dwPihtBxHVqYA+JxKx32vtiyY4Vh2infNilP/6v4mgji+LHIOgNxjl6bOsqwKHb+CLBAiC1QRDx6SQEWBFB0exObuySpTSuAgwSHZLmOcc5Yrqoyu04zbLILXodzy+Ff+SGbprJJ//+7ccF+UH0HduSwSt02eOlheh2yOOzNBCSoSlUWSOxSJ5FKfHCvOQZFoKKbF05pVEeyqEE/rh9YDcQBEMExW7lYBIJzQx8m6qootsZRxyxNfAMDtFzece0bljq7ouHrH3FapxqtOHaw6r80wOZ9EcGdTaHcw3bQrDADfXlqxty+hm0wWiUCJKtGt+yKLG4rOm+cFgM17Bp2WMVUia5inVDvWbx9Y+eJWAHjpH09bOss9duyOiZRZ6lLqSq1Tp1lVJYpAcTJrPRBRkcizH7Zf3Fh652UzuwKpNWv3iSKRKMbSRoPP8vyKU2WRvryus7kjMrfaGU8bY8W1hsnm1xY9+W774gb3tWdVVY+9ohsBklm9ssRz9HFGHmDNq3GUuuSUZlolyoenb/i1Kh9BO2q81nUtgdWvtay+fu6jtyy4aFHpu1/2pzXz5ArbiktnKiL9fM/gql+1VHmsuSdXoCiQXIkFh9o05yBSrC+z3/Fs8+wqx7waZ+6c0YaSp3S2cIZzkr54YrAqiq11PtuW9ojNYxmq2QY/OPx2lR5NZj+r3PVf/9sezxgPXD/3wkW+Cxf5Dn31ZlPPype2yZLgtom50D+VMVRZAACDMX64mZmMexxyOKnd+sQXmx++ABA5ABuhqibnqkgW1romW8hfvXr1RNPDj7YGSl3K0IdRFjCc1AUKH2zxD0S18aMqzkGkxO2Q120LftDcH4hmg/Fsx4HER9sDj7zd9vT7HS6LXFak5oJ+kWIyaw4msrt6ok27Bu2qOMx2TcZLnMreA6mO/kQwnl3bHJBFKg7xpIgQTujlbnXVNbMnlkuOEmVOUIG2dUUu++nGcpdlqGuXBIymdH8447BKZS7laJIbRBAICUQzgwlNpAQJ6AaTBVrpsQgIh9YsIQIlpCuYzGjmjFKbQHBk7IIAAiU9oVQqY1QWW6yyMHTeVyCkpTt667dqHrp5wZQ+hgAwv8a1pKF40+7Bep/tUL81g1skod5n58CPMg3kHHSTuW1SsUPmHIADktzeCm4cLkyGyarcKgAyzkeN8jiAbrIyl5KrFA6bIddNJgpw+Wllk88fjsU7/M2ymkTGGNljNvFtJByAsdz2E87YmJcfzf4LPto5BDEQzcytcZ45u2R6YF1xelljfVH3QErI9+aYvDdCYCCavXxJOUGcHliE4D9dOXMwnjULe0IaESJJvcyt3LCsMj/oj+2ySxeXX7rEt6s3KoukYGGJlOz1J689u6LEqU4nLAB4+Ob5Fon4I5m8rMfPe6MEA9GMz63cedlJeXuoj/nKKo/1seWL2vsTWcPMiyLk/Rns8Cd/fNVJnvzto56UUVx9RuU/XzNrS0eYIBYULlkkO7tj31zgXf6tGfl0F5O8fvV1c25cVrV5T4iSQjEvUSB9oYwik6dvb8yzb538T7xw55LLTyv9vC1EEKcdmEBINKl3DySevePUKo+l4GAh4hsrl96wrHJT26BmmuI06T0HEChJZPXt3ZHHly+8pLEs73+Rzy10976y/dG32+tK1LIiNTvla00lgYQT2s7e2IM3z/vR5Yd5QDMaDT/1jLZxA4bCQrnPduPfqlddOc2wAODXG7tXvrQtkWZzqhy5ksCUKXp3IOWPpn9+68LbLjxM1BMfrO1bvpz19cggU1nAbJIC2O5e6XjkQZxeWADQ7o/f89/b1m4JVHgsFW7FMPlx3UYnUGQMtndHPU7p+dsXnTffN/Tb+LqP955/niRa1OpawjgFAwWBRmJsoM/93rvyJZdOtWYNaw0++1urznzyB4ssCtm8JxRKaCIl5DhkkYSgSLE3mP6qI3zRIu8n/3bOMFJmKtWz/HYKolxXzw0jl7KDYaDTSQD0371z3Es0R9m+d0Ht1WdU/OL37b/+pPerjrDXJftcikDQHPXlHxN6FgAoQQbcH0r3h7OnVNv/9aY5151VPfLM+Lr12b2t1tqTmZYdpRopK4UCCwAcqrjq6tnfv3DGKx93vfV5f0tXDDgvccpFVolQxAmWdMjBN6nwZNb0hzMpzZhT5bjr8obbLpohi6MXZnkqyQEQCTDjz/O1iDydAQDpmm/DtGvWWO29L/vXNvs37wntC6QMHSwKdaiCVREEejD6H20lE+ccGAfN4LGUFk8bWd30FsmNM1x//Y3yq5dWiHS8+rUeHGxtmEOyaaWqimg6BSSEUVEy23bYvnOj6/WXCxfWQd+U0TfuCjZ3RJo7Ir3htD+USWRMzjklQJAcSps4B5PzXGlUIFhkEys9lplltqWzPEtnFTX47Ef5d6Ffvti1/HuqbJNLfJRQkkmYgT7LvIXFnzURq6XQYQ1t4WS2tS/eP5gNxrT+UGowoWU0rpuMAAgCUWX0OJRKt+pxSnVey+wq50TTAwZAAKKvvRH4yWrW2Un1rKQotiuuKFrzHNqO5bVReEK8mW2SjWW15PqPMRIV58yW58w95t/5/wEA3AxHCEF0JhgAAAAASUVORK5CYII="""
